+++ Scribe/Scribe-Api/src/utopia/scribe/api/database/access/many/logging/issue/ManyIssueInstancesAccess.scala
@@ -1,10 +1,14 @@
 package utopia.scribe.api.database.access.many.logging.issue
 
+import utopia.flow.generic.casting.ValueConversions._
 import utopia.scribe.api.database.access.many.logging.issue.ManyIssueInstancesAccess.ManyIssueInstancesSubView
 import utopia.scribe.api.database.factory.logging.IssueInstancesFactory
+import utopia.scribe.api.database.model.logging.IssueOccurrenceModel
 import utopia.scribe.core.model.combined.logging.IssueInstances
 import utopia.vault.nosql.factory.FromResultFactory
 import utopia.vault.sql.Condition
+
+import java.time.Instant
 
 object ManyIssueInstancesAccess
 {
@@ -23,6 +27,11 @@ object ManyIssueInstancesAccess
   */
 trait ManyIssueInstancesAccess extends ManyIssuesAccessLike[IssueInstances, ManyIssueInstancesAccess]
 {
+	// COMPUTED ----------------------------
+	
+	private def occurrenceModel = IssueOccurrenceModel
+	
+	
 	// IMPLEMENTED  ------------------------
 	
 	override protected def self: ManyIssueInstancesAccess = this
@@ -31,4 +40,13 @@ trait ManyIssueInstancesAccess extends ManyIssuesAccessLike[IssueInstances, Many
 	
 	override def filter(additionalCondition: Condition): ManyIssueInstancesAccess =
 		new ManyIssueInstancesSubView(mergeCondition(additionalCondition))
+		
+	
+	// OTHER    ----------------------------
+	
+	/**
+	  * @param threshold A time threshold
+	  * @return Access to issues that have occurred since the specified time threshold
+	  */
+	def since(threshold: Instant) = filter(occurrenceModel.latestColumn > threshold)
 }