+++ Scribe/Scribe-Core/src/utopia/scribe/core/model/combined/logging/IssueVariantInstances.scala
@@ -1,8 +1,20 @@
 package utopia.scribe.core.model.combined.logging
 
+import utopia.flow.operator.CombinedOrdering
 import utopia.flow.view.template.Extender
 import utopia.scribe.core.model.partial.logging.IssueVariantData
 import utopia.scribe.core.model.stored.logging.{IssueOccurrence, IssueVariant}
+
+object IssueVariantInstances
+{
+	/**
+	  * Ordering that orders based on software version and last occurrence time
+	  */
+	implicit val ordering: Ordering[IssueVariantInstances] = CombinedOrdering(
+		Ordering.by { v: IssueVariantInstances => v.version },
+		Ordering.by { v: IssueVariantInstances => v.occurrences.iterator.map { _.lastOccurrence }.maxOption }
+	)
+}
 
 /**
   * Adds specific occurrence/instance information to a single issue variant
@@ -18,6 +30,22 @@ case class IssueVariantInstances(variant: IssueVariant, occurrences: Vector[Issu
 	  * Id of this variant in the database
 	  */
 	def id = variant.id
+	
+	/**
+	  * @return The earliest recorded occurrence of this issue variant.
+	  *         None if no occurrences are recorded.
+	  */
+	def earliestOccurrence = occurrences.minByOption { _.firstOccurrence }
+	/**
+	  * @return The latest occurrence of this issue variant.
+	  *         None if no occurrences are recorded.
+	  */
+	def latestOccurrence = occurrences.maxByOption { _.lastOccurrence }
+	
+	/**
+	  * @return Total number of issue occurrences represented by this instance
+	  */
+	def numberOfOccurrences = occurrences.iterator.map { _.count }.sum
 	
 	
 	// IMPLEMENTED	--------------------