+++ Flow/Changes.md
@@ -19,8 +19,8 @@
   - Added `.mutate(...)`, which behaves exactly like `.pop(...)` in **Volatile**
 - **PostponingProcess** (object)
   - Added a new constructor
-- **Seq**
-  - Added `.random`
+- **Seq** (**CollectionExtensions**)
+  - Added `.reverseSorted` and `.random`
 - **Span** (object)
   - Added `.numeric(...)`
 - **SpanLike**